hilder_company/requirements.txt
aiohttp==3.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
Automat==0.7.0
beautifulsoup4==4.6.3
bs4==0.0.1
build==1.0.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cx-Oracle==6.4.1
decorator==4.2.1
Django==2.1.1
docopt==0.6.2
gevent==1.3.6
greenlet==0.4.14
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.4
multidict==4.3.1
numpy==1.15.1
pandas==0.23.4
parsel==1.5.0
pika==0.12.0
pipreqs==0.4.9
psutil==5.4.7
py==1.5.3
pyaml==17.12.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.6.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.5
pywin32==223.1
PyYAML==3.12
qiniu==7.2.2
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
retry==0.9.2
retrying==1.3.3
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
schedule==0.5.0
Scrapy==1.5.1
selenium==3.14.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.11
Twisted==18.7.0
urllib3==1.22
uuid==1.30
w3lib==1.19.0
yaml-1.3==0.1.0
yarg==0.1.9
yarl==1.2.6
zope.interface==4.5.0
